+++ backend/core/src/shared/dto/filter.dto.ts
@@ -13,5 +13,5 @@ export class BaseFilter {
     example: "=",
     default: Compare["="],
   })
-  $comparison: Compare
+  $comparison: Compare | Compare[]
 }